Using HTML, CSS, and JavaScript to create a virtual piano
One of the many things that a front-end developer has to take care of is binding the functionality of a web application to the keys on the keyboard. In order to develop and strengthen the important skills of working with keyboard keys, we will make a virtual piano in this project.